refactor: factor out pattern editor
Required to re-use the pattern editor for another component. Nice side effect: It declutters the Messages component.
