+++ source-code/website/src/pages/editor/@host/@owner/@repository/components/PatternEditor.telefunc.ts
@@ -1,4 +1,4 @@
-import { serverSideEnv } from "@env";
+import { isProduction, serverSideEnv } from "@env";
 import { assertUsage } from "@src/services/assert/index.js";
 
 const env = await serverSideEnv();
@@ -12,9 +12,9 @@ export async function onMachineTranslate(args: {
   targetLanguage: string;
 }): Promise<{ data?: string; error?: string }> {
   try {
-    if (import.meta.env.MODE === "development") {
+    if (isProduction === false) {
       throw Error(
-        "Machine translations are disabled in production due to the missing env variable GOOGLE_TRANSLATE_API_KEY. "
+        "Machine translations are disabled in development. An env variable is missing."
       );
     }
     const response = await fetch(
