--- source-code/website/src/pages/editor/@host/@owner/@repository/components/README.md
@@ -1,1 +1,0 @@
-components dedicated for the editor
